refactor : Move JwtConfig's directory path
- common/config에 포함되도록 합니다.
